whatsapp_bot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
twilio==9.3.0
redis==5.0.8
httpx==0.27.0
sarvamai==0.1.0
python-dotenv==1.0.1
python-multipart==0.0.9